Allow file permissions to be set for UNIX domain listening socket
When 'listen:' starts with a slash, it is interpreted as a UNIX domain socket
path for twemproxy to bind. This patch allows file permissions for the socket
to be specified via a space-separated permission field following the path.

Examples:

Permissions not set, so use defaults (same as before, still supported):

  listen: /var/run/nutcracker/nutcracker.sock

Set socket permissions to ugo+rw:

  listen: /var/run/nutcracker/nutcracker.sock 0666