eliminate pending kqueue messages for individual ops
Sometimes kqueue reports several events of the same type (EVFILT_WRITE)
in one batch. If we can't process them and have to close the connection's
direction (such as, disabling waiting on WRITE), we can be hit with the
same EVFILT_WRITE operation in the next iteration of the event wait loop.
This breaks the server logic, since it has assertions on (conn->send_active).

Therefore, during event_del_{in,out} operations we'v got to eliminate the
subsequent kqueue events of that particular kind from the list of events
currently being processed by the event wait loop.