Create a 0.5.0-RC1 release.
See ChangeLog for details.

There have been a lot of changes since 0.4.1 was published in 2015.

I've tested many of these on a fork in combination with the
heartbeat/failover/sentinel patches (planned for 0.6.0)
and expect them to be a significant improvement over 0.4.1,
but there may be build/runtime/packaging issues I'm not aware of on alternate
configurations or OSes.

twemproxy 0.5.0 will use github releases for distribution tarballs